requirements.txt
customtkinter
openai
python-dotenv
playwright
google-cloud-speech
sounddevice